Move things that are not "actions" from the app/actions directory to app/util for now until we come up with a better way to organize/name these bits.
