+++ assets/app/util/findShaForDefaultBranch.js
